+++ mock/postgres/init.sql
@@ -1,0 +1,9 @@
+
+CREATE DATABASE logback4js;
+\c logback4js;
+CREATE TABLE log (
+    logger text,
+    "timestamp" timestamp,
+    "level" varchar(8),
+    message text
+);